requirements.txt
flask
pandas
nltk
python-dotenv
